apps/api/pyproject.toml
[project]
name = "pactsafe-api"
version = "0.2.0"
description = "PactSafe AI FastAPI backend."
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.5.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.2.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.34.0",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "python-docx>=1.1.2",
    "reportlab>=4.2.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 110
target-version = "py311"
